[#2266] - Muda el cron de las más leídas al módulo de obras
La función leía URLs de la ruta de lectura, resolvía slugs de obra y escribía
referencias a obras, todo desde el módulo de historias y bajo un nombre que
decía "stories". Pasa a `literary-work`, con la ruta y la declaración del cron
del proveedor de hosting en este mismo commit: si el agendado quedara apuntando
a la ruta vieja, pegaría a un 404 sin que nadie se entere.

El módulo de historias conserva la lectura de las más leídas, que es lo que le
toca hasta que se dé de baja entero.

La ruta comparte prefijo con las lecturas de obra, que se sirven con caché de
borde, así que se declara incacheable: servida desde el borde devolvería un 200
sin haber corrido y el ranking quedaría congelado sin señal. El middleware pasa
a respetar un `no-store` puesto por el handler, en vez de llevar una lista de
rutas exceptuadas que hay que acordarse de actualizar.

La factory del controller recibe el repository de contenido aparte del de
obras: no lo usa ninguna lectura, es de quien escribe el ranking.
