+++ src/api/modules/literary-work/literary-work.controller.spec.ts
@@ -5,10 +5,23 @@ import { environment } from '../../_helpers/environment';
 import { readCacheHeaders } from '../../_middleware/read-cache-headers.middleware';
 import { createLiteraryWorkController } from './literary-work.controller';
 import { InMemoryLiteraryWorkRepository } from './literary-work.repository.mock';
+import { InMemoryContentRepository } from '../content/content.repository.mock';
+import { onoffLiteraryWorkNavigationTeasersWithAuthorsMock } from '@mocks/onoff-literary-work-teasers.mock';
+import { fetchClarityData } from '../../_helpers/clarity-connector';
+import type { Mock } from '@test-utils';
+
+// La ruta del cron alcanza la métrica externa por import de módulo, sin punto de inyección.
+/* eslint-disable no-restricted-syntax -- vi.mock/vi.fn: servicio externo sin seam de inyección */
+vi.mock('../../_helpers/clarity-connector', () => ({ fetchClarityData: vi.fn() }));
+/* eslint-enable no-restricted-syntax */
 
 describe('literaryWorkController', () => {
 	const controller = createLiteraryWorkController(
 		new InMemoryLiteraryWorkRepository(onoffLiteraryWorksMock, onoffLiteraryWorkTeasersMock),
+		new InMemoryContentRepository({
+			rotatingContent: { _id: 'rotatingContent', name: 'Lo más leído', mostRead: [], mostReadLiteraryWorks: [] },
+			literaryWorks: onoffLiteraryWorkNavigationTeasersWithAuthorsMock,
+		}),
 	);
 	const knownSlug = onoffLiteraryWorksMock[0].slug;
 	const [knownAuthor] = onoffLiteraryWorkTeasersMock[0].authors;
@@ -74,6 +87,19 @@ describe('literaryWorkController', () => {
 		const response = await appUnderTest().request('/literary-work/no-existe');
 
 		expect(response.headers.get('Vercel-CDN-Cache-Control')).toBeNull();
+	});
+
+	// La ruta del cron comparte prefijo con las lecturas, así que hereda su middleware de caché. Servida
+	// desde el borde devolvería un 200 sin haber corrido, y el ranking quedaría congelado sin señal.
+	it('should keep the most-read update out of the edge cache', async () => {
+		environment.production = true;
+		environment.readCacheSMaxAge = 900;
+		(fetchClarityData as Mock).mockResolvedValue([{ metricName: 'PopularPages', information: [] }]);
+
+		const response = await appUnderTest().request('/literary-work/update-most-read');
+
+		expect(response.headers.get('Vercel-CDN-Cache-Control')).toBeNull();
+		expect(response.headers.get('Cache-Control')).toBe('no-store');
 	});
 
 	describe('GET /', () => {
