+++ src/api/modules/literary-work/literary-work.controller.ts
@@ -5,9 +5,15 @@ import { slugSchema } from '@schemas/common.schemas';
 import { literaryWorkTeaserFilterSchema } from '@schemas/literary-work.schemas';
 import { LiteraryWorkNotFoundError } from './literary-work.errors';
 import type { LiteraryWorkRepository } from './literary-work.repository';
-import { getLiteraryWorkBySlug, getLiteraryWorkTeasers } from './literary-work.service';
+import type { ContentRepository } from '../content/content.repository';
+import { getLiteraryWorkBySlug, getLiteraryWorkTeasers, updateMostReadLiteraryWorks } from './literary-work.service';
 
-export function createLiteraryWorkController(repository?: LiteraryWorkRepository) {
+// El repository de contenido entra por separado porque no lo usa ninguna lectura de obras: es de quien
+// escribe el ranking, que vive acá por ser una operación sobre obras.
+export function createLiteraryWorkController(
+	repository?: LiteraryWorkRepository,
+	contentRepository?: ContentRepository,
+) {
 	const controller = new Hono();
 
 	// El catálogo, filtrable por query params: un criterio nuevo es un campo más del schema, no una
@@ -17,6 +23,17 @@ export function createLiteraryWorkController(repository?: LiteraryWorkRepository
 		const filter = c.req.valid('query');
 		const literaryWorks = await getLiteraryWorkTeasers(filter, repository);
 		return c.json(literaryWorks);
+	});
+
+	// Va antes del comodín de slug, que si no se la come tratando "update-most-read" como una obra.
+	//
+	// El `no-store` no es decorativo: el módulo sirve sus lecturas con caché de borde, y sin declararse
+	// incacheable esta escritura recibiría el mismo tratamiento y la corrida siguiente se resolvería
+	// con un hit, devolviendo un 200 sin haber actualizado nada.
+	controller.get('/update-most-read', async (c) => {
+		const result = await updateMostReadLiteraryWorks(contentRepository);
+		c.header('Cache-Control', 'no-store');
+		return c.json(result);
 	});
 
 	controller.get('/:slug', zValidator('param', slugSchema), async (c) => {
